% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cites_connect.R
\name{cites_connect}
\alias{cites_connect}
\title{Creates a connection.}
\usage{

  cites_connect(url = "https://api.speciesplus.net/api/v1/taxon_concepts.xml?name=",
  token)
}
\arguments{
\item{url}{The base URL to connect to. Default value:}

\item{token}{The user token}
}
\value{
connect returns a connection object for CITES.
This connection object can be used to discover and access to resources.
}
\description{
Setup a connection to CITES with API token.
}
\examples{
cnx <- cites_connect(token = 'ErJcYxUsIApHLCLOxiJ1Zwtt')
}
