% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cites_query.R
\name{cites_query}
\alias{cites_query}
\title{Access data...}
\usage{
cites_query(cnx, query = "Smaug giganteus", appendix_only = TRUE)
}
\arguments{
\item{cnx}{the output of your cites_connect() function.}

\item{query}{a query... character}

\item{appendix_only}{a logical}
}
\value{
xml so far...
}
\description{
Query CITES.
}
\examples{
cnx <- cites_connect(token = 'ErJcYxUsIApHLCLOxiJ1Zwtt')
# cites_query(cnx, query = 'Smaug giganteus', appendix_only = TRUE)
cites_query(cnx, query = 'Homo sapiens', appendix_only = TRUE)
}
