% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{pack}
\name{citesr}
\alias{citesr}
\alias{spplus_baseurl}
\title{citesr:}
\usage{
spplus_baseurl()
}
\description{
An R packe to access the CITES Species+ database (https://speciesplus.net/)
}
\author{
Jonas Geschke et al.
}
