% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helloWorld.R
\name{helloWord}
\alias{helloWord}
\alias{helloWorld}
\title{A first funciton}
\usage{
helloWorld(name = "world")
}
\arguments{
\item{name}{a charcater string.}
}
\value{
Print a message
}
\description{
example function
}
\details{
Add valuable details
}
\examples{
helloWorld('Ionas')

}
