% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spp_cites_legislation.R
\name{spp_cites_legislation}
\alias{spp_cites_legislation}
\title{Get current CITES appendix listings and reservations.}
\usage{
spp_cites_legislation(taxon_id, scope = "current", language = "en",
  raw = FALSE, token = NULL)
}
\arguments{
\item{taxon_id}{character string containing a species' taxon concept identifier
(see \code{\link[=spp_taxonconcept]{spp_taxonconcept()}}).}

\item{scope}{vector of character strings indicating the time scope of legislation,
values are taken among \code{current}, \code{historic} and \code{all}.
Default is \code{current}.}

\item{language}{vector of character strings indicating the language for the
text of legislation notes, values are taken among \code{en} (English),
\code{fr} (French) and \code{es} (Spanish). Default is \code{en}.}

\item{raw}{a logical. Should raw data be returned?}

\item{token}{a character string containing the authentification token, see
\url{https://api.speciesplus.net/documentation}. Default is set to
\code{NULL} and requires the environment variable \code{SPECIESPLUS_TOKEN} to be
set directly in \code{Renviron}. Alternatively, \code{\link[=set_token]{set_token()}} can
be used to set \code{SPECIESPLUS_TOKEN} for the current session.}
}
\value{
An object of class `` that is a list of three data frames:
}
\description{
Retrieve current CITES appendix listings and reservations, CITES quotas, and
CITES suspensions for a given taxon concept.
}
\examples{
\donttest{
res1 <- spp_cites_legislation(taxon_id = 4521)
res2 <- spp_cites_legislation(taxon_id = 4521, scope = 'all')
res3 <- spp_cites_legislation(taxon_id = 4521, language = 'fr')
}
}
\references{
\url{https://api.speciesplus.net/documentation/v1/cites_legislation/index.html}
}
