% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spp_distributions.R
\name{spp_distributions}
\alias{spp_distributions}
\title{Get distributions data available for a given taxon concept.}
\usage{
spp_distributions(taxon_id, language = "en", raw = FALSE,
  token = NULL)
}
\arguments{
\item{taxon_id}{character string containing a species' taxon concept identifier
(see \code{\link[rcites]{spp_taxonconcept}}).}

\item{language}{vector of character strings indicating the language for the
names of distributions, values are taken among \code{en} (English),
\code{fr} (French) and \code{es} (Spanish). Default is \code{en}.}

\item{raw}{a logical. Should raw data be returned?}

\item{token}{a character string containing the authentification token, see
\url{https://api.speciesplus.net/documentation}. Default is set to
\code{NULL} and requires the environment variable \code{SPECIESPLUS_TOKEN} to be
set directly in \code{Renviron}. Alternatively \code{set_token()} can
be used to set \code{SPECIESPLUS_TOKEN} for the current session.}
}
\value{
A data frame with all distribution information.
}
\description{
Retrieve distributions data available for a given taxon concept for which the
the taxon identifier is known.
}
\examples{
\donttest{
 res1 <- spp_distributions(taxon_id = '4521')
 res2 <- spp_distributions(taxon_id = '4521', raw = TRUE)
 res3 <- spp_distributions(taxon_id = '4521', language = 'fr')
}
}
\references{
\url{https://api.speciesplus.net/documentation/v1/distributions/index.html}
}
