% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spp_references.R
\name{spp_references}
\alias{spp_references}
\title{Get references for a given taxon concept}
\usage{
spp_references(taxon_id, type = c("taxonomic", "distribution"),
  simplify = FALSE, token = NULL)
}
\arguments{
\item{taxon_id}{character string containing a species' taxon concept identifier
(see \code{\link[rcites]{spp_taxonconcept}}).}

\item{type}{vector of character strings indicating the type of references
requested, \code{taxonomic} or \code{distribution}.}

\item{simplify}{a logical. Should the output be simplified? In other words,
should columns of data.table objects returned be unlisted when they are
lists made of single elements?}

\item{token}{a character string containing the authentification token, see
\url{https://api.speciesplus.net/documentation}. Default is set to
\code{NULL} and requires the environment variable \code{SPECIESPLUS_TOKEN} to be
set directly in \code{Renviron}. Alternatively \code{set_token()} can
be used to set \code{SPECIESPLUS_TOKEN} for the current session.}
}
\value{
A list of data.table objects, one per type requested.
}
\description{
Retrieve available references for a given taxon concept.
}
\examples{
\donttest{
res1 <- spp_references(taxon_id = '4521')
res2 <- spp_references(taxon_id = '4521', type = 'taxonomic', simplify = T)
}
}
\references{
\url{https://api.speciesplus.net/documentation/v1/references/index.html}
}
