% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sppplus_connect.R
\name{sppplus_connect}
\alias{sppplus_connect}
\title{Create CITES species+ connection details}
\usage{
sppplus_connect(url = "https://api.speciesplus.net/api/v1/", token)
}
\arguments{
\item{url}{The base URL to connect to. Default value is the Taxon connect API: \url{\https://api.speciesplus.net/api/v1/taxon_concepts.xml?name=}}

\item{token}{Your species+ API authentication token}
}
\value{
Returns a list of credentials for accessing CITES species+ API. The list contains the URL and token.
}
\description{
Creates the connection for future CITES species+ query.
Generate your own API token from Species+ API dashboard \url{https://api.speciesplus.net}
}
\examples{
# Not run: #
# cnx <- sppplus_connect(token = 'insert your token here')
}
