% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sppplus_login.R
\name{sppplus_login}
\alias{sppplus_login}
\title{Login helper function}
\usage{
sppplus_login(token = NULL)
}
\arguments{
\item{token}{a character string containing you token. If `NULL` then the
token can be passed interactively.}
}
\description{
This function sets the authentification token for the current session.
}
\examples{
\donttest{
 sppplus_login()
}
}
\references{
\url{https://api.speciesplus.net/documentation}
}
