% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sppplus_simplify.R
\name{sppplus_simplify}
\alias{sppplus_simplify}
\title{Simplify outputs}
\usage{
sppplus_simplify(x)
}
\arguments{
\item{x}{a data.table to be simplified.}
}
\value{
A data table with a simplified structure.
}
\description{
Simplify the structure of cites/species+ API outputs.
}
\details{
For the sake of generality, CITES Species+ API outputs are parsed
and stored in data.table objects whose columns are lists. Most of the lists
are actually well-structured and can actually be stored as columns of integer,
date, characters strings, etc. This function handles such simplifications.
}
