% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon_cites_legislation.R
\name{taxon_cites_legislation}
\alias{taxon_cites_legislation}
\title{Access distribution data from CITES species+ API}
\usage{
taxon_cites_legislation(token, tax_id)
}
\arguments{
\item{token}{Authentification token, see \url{https://api.speciesplus.net/documentation}.}

\item{tax_id}{character string containing a species' taxon id (e.g. 4521), which is returned by \code{\link[citesr]{sppplus_taxonconcept}}.}
}
\value{
If type is one of listing, quota or suspension, returns a dataframe with respective information from the Species+ database. If type is all, returns a list with all Species+ database information.
}
\description{
Queries CITES species+ API using an authentication token.
The query string filters species+ data by taxon concept (e.g.species, genus, class)
}
\examples{
# taxon_cites_legislation(token, tax_id = '4521')
}
\references{
\url{https://api.speciesplus.net/documentation/v1/cites_legislation/index.html}
}
