% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon_distribution.R
\name{taxon_distribution}
\alias{taxon_distribution}
\title{Access distribution data from CITES species+ API}
\usage{
taxon_distribution(token, tax_id)
}
\arguments{
\item{token}{Authentification token, see \url{https://api.speciesplus.net/documentation}.}

\item{tax_id}{character string containing a species' taxon id (e.g. 4521), which is returned by \code{\link[citesr]{sppplus_taxonconcept}}.}
}
\value{
If country_only is \code{TRUE}, returns a dataframe with taxon and country. If appendix_only is FALSE, returns a list with all distribution information.
}
\description{
Queries CITES species+ API using an authentication token. The query string
filters species+ data by taxon concept (e.g. species, genus, class).
}
\examples{
# taxon_distribution(token, tax_id = '4521', country_only = TRUE)
}
\references{
\url{https://api.speciesplus.net/documentation/v1/distributions/index.html}
}
