% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{citesr}
\alias{citesr}
\alias{sppplus_baseurl}
\alias{citesr-package}
\title{citesr:}
\usage{
sppplus_baseurl()
}
\description{
An R package to access the CITES Species+ database (https://speciesplus.net/)
}
\author{
Jonas Geschke et al.
}
