% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.spp.R
\name{print.spp}
\alias{print.spp}
\alias{print.spp_raw}
\alias{print.spp_raw_multi}
\alias{print.spp_cites_leg}
\alias{print.spp_cites_leg_multi}
\alias{print.spp_distr}
\alias{print.spp_distr_multi}
\alias{print.spp_eu_leg}
\alias{print.spp_eu_leg_multi}
\alias{print.spp_refs}
\alias{print.spp_refs_multi}
\alias{print.spp_taxon}
\title{Print methods for objects of class \verb{spp_raw*}.}
\usage{
\method{print}{spp_raw}(x, ...)

\method{print}{spp_raw_multi}(x, ...)

\method{print}{spp_cites_leg}(x, ...)

\method{print}{spp_cites_leg_multi}(x, ...)

\method{print}{spp_distr}(x, ...)

\method{print}{spp_distr_multi}(x, ...)

\method{print}{spp_eu_leg}(x, ...)

\method{print}{spp_eu_leg_multi}(x, ...)

\method{print}{spp_refs}(x, ...)

\method{print}{spp_refs_multi}(x, ...)

\method{print}{spp_taxon}(x, ...)
}
\arguments{
\item{x}{an object of class \verb{spp_raw*}.}

\item{...}{ignored.}
}
\value{
The JSON result.
}
\description{
Print the outputs of a Species+ API call.
}
