% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{rcites}
\alias{rcites}
\alias{sppplus_baseurl}
\alias{rcites-package}
\title{rcites:}
\usage{
sppplus_baseurl()
}
\description{
A programmatic interface to the Species+ <https://speciesplus.net/> database
via the Species+/CITES Checklist API <https://api.speciesplus.net/>.
}
