% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_token.R
\name{set_token}
\alias{set_token}
\title{Login helper function}
\usage{
set_token(token = NULL)
}
\arguments{
\item{token}{a character string containing you token. If `NULL` then the
token can be passed interactively.}
}
\description{
This function sets the authentification token for the current session.
}
\examples{
\donttest{
 set_token()
}
}
\references{
\url{https://api.speciesplus.net/documentation}
}
