% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_token.R
\name{set_token}
\alias{set_token}
\title{Login helper function}
\usage{
set_token(token = NULL)
}
\arguments{
\item{token}{a character string (with quotes) containing you token. If `NULL` then the
token can be passed without quotes (not as character string) after a prompt.}
}
\description{
This function sets the authentification token for the current session.
}
\examples{
\donttest{
 # NB the token below is not working
 set_token('8QW6Qgh57sBG2k0gtt')
 # interactively
 set_token()
}
}
\references{
\url{https://api.speciesplus.net/documentation}
}
