% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_token.R
\name{set_token}
\alias{set_token}
\alias{forget_token}
\title{Login helper function}
\usage{
set_token(token = NULL)

forget_token()
}
\arguments{
\item{token}{a character string (with quotes) containing your token. If
\code{NULL}, then the token can be passed without quotes (not as character
string) after a prompt.}
}
\description{
Set and forget the authentification token for the current session.
}
\section{Functions}{
\itemize{
\item \code{set_token}: set the environment variable \code{SPECIESPLUS_TOKEN}.

\item \code{forget_token}: forget the environment variable \code{SPECIESPLUS_TOKEN}.
}}

\examples{
\dontrun{
 # NB: the token below is not working
 set_token('8QW6Qgh57sBG2k0gtt')
 # interactively
 set_token()
}
}
\references{
\url{https://api.speciesplus.net/documentation}
}
