% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spp_cites_legislation.R
\name{spp_cites_legislation}
\alias{spp_cites_legislation}
\title{Get current CITES appendix listings and reservations}
\usage{
spp_cites_legislation(
  taxon_id,
  scope = "current",
  language = "en",
  raw = FALSE,
  token = NULL,
  verbose = TRUE,
  pause = 1,
  ...
)
}
\arguments{
\item{taxon_id}{a vector of character strings containing species' taxon
concept identifiers (see \code{\link[=spp_taxonconcept]{spp_taxonconcept()}}).}

\item{scope}{vector of character strings indicating the time scope of
legislation, values are taken among \code{current}, \code{historic} and \code{all}. Default
is \code{current}.}

\item{language}{vector of character strings indicating the language for the
text of legislation notes, values are taken among \code{en} (English),
\code{fr} (French) and \code{es} (Spanish). Default is \code{en}.}

\item{raw}{a logical. Should raw data be returned?}

\item{token}{a character string containing the authentification token, see
\url{https://api.speciesplus.net/documentation}. Default is set to
\code{NULL} and requires the environment variable \code{SPECIESPLUS_TOKEN} to be
set directly in \code{Renviron}. Alternatively, \code{\link[=set_token]{set_token()}} can
be used to set \code{SPECIESPLUS_TOKEN} for the current session.}

\item{verbose}{a logical. Should extra information be reported on progress?}

\item{pause}{a duration (in second) to suspend execution for (see
\code{\link[=Sys.sleep]{Sys.sleep()}}). This was added cause the web API returns a 404 error too many
requests in a short time interval.}

\item{...}{Further named parameters, see \code{\link[httr:GET]{httr::GET()}}.}
}
\value{
If \code{raw} is set to \code{TRUE} then an object of class \code{spp_raw} (or
\code{spp_raw_multi} if \code{length(taxon_id) > 1}) is returned which is essentially
a list of lists (see option \code{as = 'parsed'} in \code{\link[httr:content]{httr::content()}}).
Otherwise, an object of class \code{spp_cites_leg} (or \code{spp_cites_leg_multi} if
\code{length(taxon_id)>1}) is returned which is a list of three data frames:
\enumerate{
\item \code{cites_listings}: lists CITES annex listings EU suspensions,
\item \code{cites_quotas}: lists CITES quotas,
\item \code{cites_suspensions}: lists CITES suspensions.
}
}
\description{
Retrieve current CITES appendix listings and reservations, CITES quotas, and
CITES suspensions for a given taxon concept.
}
\examples{
\dontrun{
# this calls will only work if a token is set and valid
res1 <- spp_cites_legislation(taxon_id = 4521)
res2 <- spp_cites_legislation(taxon_id = c('4521', '3210', '10255'))
res3 <- spp_cites_legislation(taxon_id = 4521, scope = 'all',
verbose = FALSE, config=httr::verbose())
res4 <- spp_cites_legislation(taxon_id = 4521, language = 'fr')
}
}
\references{
\url{https://api.speciesplus.net/documentation/v1/cites_legislation/index.html}
}
