% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spp_references.R
\name{spp_references}
\alias{spp_references}
\title{Get references for a given taxon concept}
\usage{
spp_references(taxon_id, raw = FALSE, token = NULL)
}
\arguments{
\item{taxon_id}{character string containing a species' taxon concept identifier
(see \code{\link[rcites]{spp_taxonconcept}}).}

\item{raw}{a logical. Should raw data be returned?}

\item{token}{a character string containing the authentification token, see
\url{https://api.speciesplus.net/documentation}. Default is set to
\code{NULL} and requires the environment variable \code{SPECIESPLUS_TOKEN} to be
set directly in \code{Renviron}. Alternatively, \code{\link[=set_token]{set_token()}} can be used to set
\code{SPECIESPLUS_TOKEN} for the current session.}
}
\value{
If \code{raw} is set to \code{TRUE} then an object of class \code{spp_raw} is returned
which is essentially the list of lists (see option \code{as = 'parsed'} in \code{\link[httr:content]{httr::content()}}).
Otherwise an object of class \code{spp_refs} is returned which is a list of one
data frame:
\itemize{
\item \code{references} that includes the identifier of the reference and the
corresponding citation.
}
}
\description{
Retrieve available references for a given taxon concept.
}
\examples{
\donttest{
res1 <- spp_references(taxon_id = '4521')
res2 <- spp_references(taxon_id = '3210')
res3 <- spp_references(taxon_id = '4521', raw = TRUE)
}
}
\references{
\url{https://api.speciesplus.net/documentation/v1/references/index.html}
}
