% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spp_references.R
\name{spp_references}
\alias{spp_references}
\title{Get references for a given taxon concept}
\usage{
spp_references(taxon_id, raw = FALSE, token = NULL)
}
\arguments{
\item{taxon_id}{character string containing a species' taxon concept identifier
(see \code{\link[rcites]{spp_taxonconcept}}).}

\item{raw}{a logical. Should raw data be returned?}

\item{token}{a character string containing the authentification token, see
\url{https://api.speciesplus.net/documentation}. Default is set to
\code{NULL} and requires the environment variable \code{SPECIESPLUS_TOKEN} to be
set directly in \code{Renviron}. Alternatively \code{set_token()} can
be used to set \code{SPECIESPLUS_TOKEN} for the current session.}
}
\value{
A list of objects, one per type requested.
}
\description{
Retrieve available references for a given taxon concept.
}
\examples{
\donttest{
res1 <- spp_references(taxon_id = '4521')
}
}
\references{
\url{https://api.speciesplus.net/documentation/v1/references/index.html}
}
