% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spp_taxonconcept.R
\name{spp_taxonconcept}
\alias{spp_taxonconcept}
\title{Get taxon concepts for a search term.}
\usage{
spp_taxonconcept(query_taxon, taxonomy = "CITES",
  with_descendants = FALSE, language = NULL, updated_since = NULL,
  per_page = 500, seq_page = NULL, raw = FALSE, token = NULL,
  verbose = TRUE)
}
\arguments{
\item{query_taxon}{a character string containing the query (e.g. species).
Scientific taxa only (max 255 characters).}

\item{taxonomy}{filter taxon concepts by taxonomy, accepts either 'CITES' or
'CMS' as its value. Defaults set to 'CITES'.}

\item{with_descendants}{a logical. Should the search by name be broadened to
include higher taxa?}

\item{language}{filter languages returned for common names. Value should be a
vector of sting including one or more country codes. Default is set to `NULL`,
showing all available languages.}

\item{updated_since}{a timestamp. Only objects updated after (and including)
this timestamp will be pulled.}

\item{per_page}{a integer that indicates how many objects are returned per
page for paginated responses. Default set to 500 which is the maximum.}

\item{seq_page}{a vector of integer that contains page numbers. Default is
set to `NULL`, i.e. all pages are accessed.}

\item{raw}{a logical. Should raw data be returned?}

\item{token}{a character string containing the authentification token, see
\url{https://api.speciesplus.net/documentation}. Default is set to
\code{NULL} and requires the environment variable \code{SPECIESPLUS_TOKEN} to be
set directly in \code{Renviron}. Alternatively, \code{set_token()} can
be used to set \code{SPECIESPLUS_TOKEN} for the current session.}

\item{verbose}{a logical. Should extra information be reported on progress?}
}
\value{
If `raw=TRUE`, then a object of class `spp_raw` is returned, which is essentially
a list of lists. If `raw=FALSE`, then a object of class `spp_taxon is returned
it is a collection of five data frames:
1. general
2. higher_taxa
3.
4. accepted_names
5.results If then `NULL` is returned with a warning message. Othe
}
\description{
Retrieve the taxon concept of a specific taxon (scientific name).
}
\examples{
\donttest{
res1 <- spp_taxonconcept(query_taxon = 'Loxodonta africana')
res2 <- spp_taxonconcept(query_taxon = 'Loxodonta africana', raw = TRUE)
res3 <- spp_taxonconcept(query_taxon = 'Amazilia versicolor')
res4 <- spp_taxonconcept(query_taxon = '', taxonomy = 'CMS', seq_page = 1:5, language = 'EN')
res5 <- spp_taxonconcept(query_taxon = '', seq_page = c(44))
}
}
\references{
\url{https://api.speciesplus.net/documentation/v1/taxon_concepts/index.html}
}
