% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cites_connect.R
\name{sppplus_connect}
\alias{sppplus_connect}
\title{Create CITES species+ connection details}
\usage{

  sppplus_connect(url = "https://api.speciesplus.net/api/v1/taxon_concepts.xml?name=",
  token)
}
\arguments{
\item{url}{The base URL to connect to. Default value is the Taxon connect API: https://api.speciesplus.net/api/v1/taxon_concepts.xml?name=}

\item{token}{The species+ API authentication token}
}
\value{
Returns a list of credentials for accessing CITES species+ API. The list contains the URL and token.
}
\description{
Creates the connection information for CITES species+ query. Takes the API token generated from Species+ API dashboard \url{https://api.speciesplus.net}
}
\examples{
cnx <- sppplus_connect(token = 'ErJcYxUsIApHLCLOxiJ1Zwtt')
}
