% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sppplus_taxonconcept.R
\name{sppplus_taxonconcept}
\alias{sppplus_taxonconcept}
\title{Access taxon_concept data from CITES species+ API}
\usage{
sppplus_taxonconcept(token, query_taxon, appendix_only = TRUE)
}
\arguments{
\item{token}{Authentification token, see \url{https://api.speciesplus.net/documentation}.}

\item{query_taxon}{character string containing the query (e.g. species). Scientific taxa only.}

\item{appendix_only}{logical statement for querying only the taxon and CITES appendix information. Default is set to \code{TRUE}.}
}
\value{
If appendix_only is \code{TRUE}, returns a dataframe with a species' taxon id and CITES appendix information. If appendix_only is \code{FALSE}, returns a list with all Species+ taxon concept information.
}
\description{
Queries CITES species+ API using an authentication token.
The query string filters species+ data by taxon concept (e.g.species, genus, class)
}
\examples{
# sppplus_taxonconcept(token = token, query_taxon = 'Loxodonta africana', appendix_only = TRUE)
# sppplus_taxonconcept(token = token, query_taxon = 'Homo sapiens', appendix_only = TRUE)
}
\references{
\url{https://api.speciesplus.net/documentation/v1/taxon_concepts/index.html}
}
