% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sppplus_taxonconcept.R
\name{sppplus_taxonconcept}
\alias{sppplus_taxonconcept}
\title{Access taxon_concept data from CITES species+ API}
\usage{
sppplus_taxonconcept(cnx, query = "Smaug giganteus", appendix_only = TRUE)
}
\arguments{
\item{cnx}{species+ connection information (see \code{\link[citesr]{sppplus_connect}})}

\item{query}{character string containing the query}

\item{appendix_only}{a logical statement for querying information in the Appendix. Default is TRUE.}
}
\value{
If appendix_only is TRUE, returns a dataframe with taxon and cites_listing. If appendix_only is FALSE, returns a list with all matching taxon_concepts.
}
\description{
Queries CITES species+ API using connection generated from \code{\link[citesr]{sppplus_connect}}. The query string filters species+ data by taxon concept (e.g.species, genus, class)
}
\examples{
cnx <- sppplus_connect(token = 'ErJcYxUsIApHLCLOxiJ1Zwtt')
sppplus_taxonconcept(cnx, query = 'Smaug giganteus', appendix_only = TRUE)
sppplus_taxonconcept(cnx, query = 'Homo sapiens', appendix_only = TRUE)
}
