% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon_cites_legislation.R
\name{taxon_cites_legislation}
\alias{taxon_cites_legislation}
\title{Access CITES legislation data from CITES species+ API}
\usage{
taxon_cites_legislation(cnx, query_taxon = "Loxodonta africana",
  tax_id = NULL, type = "listing")
}
\arguments{
\item{cnx}{the output of your \code{sppplus_connect()} function.}

\item{query_taxon}{a character vector to be documented}

\item{tax_id}{to be documented.}

\item{type}{to be documented.}
}
\value{
data frame or list...
}
\description{
Query CITES.
}
\examples{
# cnx <- sppplus_connect(token = 'ErJcYxUsIApHLCLOxiJ1Zwtt')
# taxon_cites_legislation(cnx, tax_id = '4521', type = 'listing')
}
