% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon_cites_legislation.R
\name{taxon_cites_legislation}
\alias{taxon_cites_legislation}
\title{Access distribution data from CITES species+ API}
\usage{
taxon_cites_legislation(cnx, tax_id = "4521", type = "listing")
}
\arguments{
\item{cnx}{species+ connection information (see \code{\link[citesr]{sppplus_connect}}).}

\item{tax_id}{character string containing the taxon id (e.g. 4521), which is returned by \code{\link[citesr]{sppplus_taxonconcept}}.}

\item{type}{character string indicating type of legislation information requested. One of listing, quota, suspension and all. Default is listing.}
}
\value{
If type is one of listing, quota or suspension, returns a dataframe with respective information from the Species+ database. If type is all, returns a list with all Species+ database information.
}
\description{
Queries CITES species+ API using connection generated from \code{\link[citesr]{sppplus_connect}}. The query string filters species+ data by taxon concept (e.g.species, genus, class)
}
\examples{
# cnx <- sppplus_connect(token = 'insert your token here')
# taxon_cites_legislation(cnx, tax_id = '4521', type = 'listing')
}
