% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon_cites_legislation.R
\name{taxon_cites_legislation}
\alias{taxon_cites_legislation}
\title{Access legislation data from CITES species+ API}
\usage{
taxon_cites_legislation(tax_id, token = NULL, type = c("listings", "quotas",
  "suspensions"))
}
\arguments{
\item{tax_id}{character string containing a species' taxon id (e.g. 4521), which is returned by \code{\link[citesr]{sppplus_taxonconcept}}.}

\item{token}{Authentification token, see \url{https://api.speciesplus.net/documentation}. Default is set to \code{NULL} and require the environment variable \code{SPPPLUS_TOKEN} to be set directly in \code{.Renviron} or for the session using \code{sppplus_login()}.}

\item{type}{vector of character strings indicating type of legislation information requested, values are taken among \code{listing}, \code{quota} and \code{suspension}. Default includes the three of them.}
}
\value{
A list of data.table objects, one per type requested.
}
\description{
Queries CITES species+ API using an authentication token.
The query string filters species+ data by taxon concept (e.g.species, genus, class)
}
\examples{
# taxon_cites_legislation(token, tax_id = '4521')
# taxon_cites_legislation(token, tax_id = '4521', type ='listings')
}
\references{
\url{https://api.speciesplus.net/documentation/v1/cites_legislation/index.html}
}
