% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon_cites_legislation.R
\name{taxon_cites_legislation}
\alias{taxon_cites_legislation}
\title{Access distribution data}
\usage{
taxon_cites_legislation(cnx, query_taxon = "Loxodonta africana",
  suspension = TRUE)
}
\arguments{
\item{cnx}{the output of your sppplus_connect() function.}

\item{query}{a query... character}
}
\value{
data frame or list...
}
\description{
Query CITES.
}
\examples{
cnx <- sppplus_connect(token = 'ErJcYxUsIApHLCLOxiJ1Zwtt')
taxon_cites_legislation(cnx, query_taxon = 'Loxodonta africana', suspension = TRUE)
}
