% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon_distribution.R
\name{taxon_distribution}
\alias{taxon_distribution}
\title{Access distribution data from CITES species+ API}
\usage{
taxon_distribution(cnx, tax_id = "4521", country_only = TRUE)
}
\arguments{
\item{cnx}{species+ connection information (see \code{\link[citesr]{sppplus_connect}}).}

\item{tax_id}{character string containing the taxon id (e.g. 4521), which is returned by \code{\link[citesr]{sppplus_taxonconcept}}.}

\item{country_only}{logical statement for querying only the country information. Default is TRUE.}
}
\value{
If country_only is TRUE, returns a dataframe with taxon and country. If appendix_only is FALSE, returns a list with all distribution information.
}
\description{
Queries CITES species+ API using connection generated from \code{\link[citesr]{sppplus_connect}}. The query string filters species+ data by taxon concept (e.g.species, genus, class)
}
\examples{
# cnx <- sppplus_connect(token = 'insert your token here')
# taxon_distribution(cnx, tax_id = '4521', country_only = TRUE)
}
