% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon_distribution.R
\name{taxon_distribution}
\alias{taxon_distribution}
\title{Access distribution data from CITES species+ API}
\usage{
taxon_distribution(cnx, query_taxon = "Loxodonta africana", tax_id = NULL,
  country_only = TRUE)
}
\arguments{
\item{tax_id}{character string containing the tax id (e.g. 4521)}

\item{country_only}{a logical statement for querying only the country information. Default is TRUE.}

\item{query}{character string containing the query (e.g. species). Depreciated if tax_id is provided.}
}
\value{
If country_only is TRUE, returns a dataframe with taxon and country. If appendix_only is FALSE, returns a list with all distribution information.
}
\description{
Queries CITES species+ API using connection generated from \code{\link[citesr]{sppplus_connect}}. The query string filters species+ data by taxon concept (e.g.species, genus, class)
}
\examples{
cnx <- sppplus_connect(token = 'nsert your token here')
taxon_distribution(cnx, query_taxon = 'Loxodonta africana', country_only = TRUE)
}
