% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon_references.R
\name{taxon_references}
\alias{taxon_references}
\title{Access distribution data from CITES species+ API}
\usage{
taxon_references(tax_id, type = c("taxonomic", "distribution"),
  simplify = FALSE, token = NULL)
}
\arguments{
\item{tax_id}{character string containing a species' taxon id (e.g. 4521), which is returned by \code{\link[citesr]{sppplus_taxonconcept}}.}

\item{type}{vector of character strings indicating the type of references requested, \code{taxonomic} or \code{distribution}.}

\item{simplify}{a logical. Should the output be simplified? In other words should columns of data.table objects be unlisted when possible? Default is set to \code{FALSE}.}

\item{token}{Authentification token, see \url{https://api.speciesplus.net/documentation}. Default is set to \code{NULL} and require the environment variable \code{SPPPLUS_TOKEN} to be set directly in \code{.Renviron} or for the session using \code{sppplus_login()}.}
}
\value{
A list of data.table objects, one per type requested.
}
\description{
Queries CITES species+ API using an authentication token. The query string
filters species+ data by taxon concept (e.g. species, genus, class)
}
\examples{
# not run:
# res1 <- taxon_references(tax_id = '4521')
# res2 <- taxon_references(tax_id = '4521', type = 'taxonomic', simplify = T)
}
\references{
\url{https://api.speciesplus.net/documentation/v1/references/index.html}
}
