% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon_references.R
\name{taxon_references}
\alias{taxon_references}
\title{Access distribution data from CITES species+ API}
\usage{
taxon_references(token, tax_id = "4521")
}
\arguments{
\item{token}{Authentification token, see \url{https://api.speciesplus.net/documentation}.}

\item{tax_id}{character string containing a species' taxon id (e.g. 4521), which is returned by \code{\link[citesr]{sppplus_taxonconcept}}.}
}
\value{
If type is one of taxonomic or distribution, returns a dataframe with respective references from the Species+ database. If type is distribution, check \code{\link[citesr]{taxon_distribution}} for country notes.
}
\description{
Queries CITES species+ API using an authentication token. The query string
filters species+ data by taxon concept (e.g. species, genus, class)
}
\examples{
# taxon_references(token, tax_id = '4521')
}
\references{
\url{https://api.speciesplus.net/documentation/v1/references/index.html}
}
