% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon_references.R
\name{taxon_references}
\alias{taxon_references}
\title{Access distribution data from CITES species+ API}
\usage{
taxon_references(token, tax_id = "4521", type = c("taxonomic",
  "distribution"), simplify = FALSE)
}
\arguments{
\item{token}{Authentification token, see \url{https://api.speciesplus.net/documentation}.}

\item{tax_id}{character string containing a species' taxon id (e.g. 4521), which is returned by \code{\link[citesr]{sppplus_taxonconcept}}.}

\item{type}{vector of character strings indicating the type of references requested, \code{taxonomic} or \code{distribution}.}

\item{simplify}{a logical. Should the output be simplified? In other words should columns of data.table objects be unlisted when possible? Default is set to \code{FALSE}.}
}
\value{
A list of data table with the desired references.
}
\description{
Queries CITES species+ API using an authentication token. The query string
filters species+ data by taxon concept (e.g. species, genus, class)
}
\examples{
# taxon_references(token, tax_id = '4521')
# taxon_references(token, tax_id = '4521', type = 'taxonomic')
}
\references{
\url{https://api.speciesplus.net/documentation/v1/references/index.html}
}
